refact(user-nodes-stats): hooks and transformations
