+++ dashboard/src/modules/users/components/widgets/user-nodes-usage-widget/hooks/use-chart-config.ts
@@ -1,0 +1,25 @@
+import {
+    ChartConfig,
+} from "@marzneshin/components";
+import { UserNodeUsagesResponse } from "@marzneshin/modules/users";
+import { interpolateColors } from "@marzneshin/utils";
+import { interpolateRainbow } from "d3";
+
+export const useChartConfig = (nodesUsage: UserNodeUsagesResponse) => {
+    const numberOfNodes = nodesUsage.node_usages.length;
+    const config: Record<string, any> = {
+        views: {
+            label: "Page Views",
+        }
+    }
+    const colorRangeInfo = {
+        colorStart: 0,
+        colorEnd: 1,
+        useEndAsStart: false,
+    };
+    const colors = interpolateColors(numberOfNodes, interpolateRainbow, colorRangeInfo);
+    nodesUsage.node_usages.forEach((node, i) => {
+        config[node.node_name] = { label: node.node_name, color: colors[i] };
+    })
+    return config satisfies ChartConfig;
+}
