+++ dashboard/src/modules/users/components/widgets/user-nodes-usage-widget/index.tsx
@@ -1,67 +1,19 @@
 import { FC } from 'react';
-import { transformData } from "./transform-data";
 import { useTranslation } from 'react-i18next';
 import {
     SectionWidget,
     ChartContainer,
     ChartTooltip,
     ChartTooltipContent,
-    ChartConfig,
 } from "@marzneshin/components";
-import { UserType, UserNodeUsagesResponse, useUserNodeUsagesQuery } from "@marzneshin/modules/users";
+import { UserNodeUsagesResponse, useUserNodeUsagesQuery } from "@marzneshin/modules/users";
 import mockData from "./mock-data.json";
-import { Area, AreaChart, CartesianGrid, XAxis } from "recharts"
+import { Area, AreaChart, CartesianGrid, YAxis, XAxis } from "recharts"
 import { useState } from "react";
 import { SelectDateView } from "./select-date-view";
-import { interpolateColors } from "@marzneshin/utils";
-import { interpolateRainbow } from "d3";
+import { UserNodesUsageWidgetProps } from "./types";
+import { useChartConfig, useTransformData } from "./hooks";
 
-export interface NodesUsage {
-    datetime: Date;
-    nodes: Record<string, number>;
-}
-
-interface UserNodesUsageWidgetProps {
-    user: UserType;
-}
-
-type ChartDateInterval = '90d' | '30d' | '7d';
-
-function useStartEndDateFromNow(timeRange: ChartDateInterval) {
-    const start = new Date();
-    const end = new Date();
-    let daysToSubtract = 90;
-    if (timeRange === "30d") {
-        daysToSubtract = 30;
-    } else if (timeRange === "7d") {
-        daysToSubtract = 7;
-    }
-    start.setDate(end.getDate() - daysToSubtract);
-    return {
-        start: start.toISOString(),
-        end: end.toISOString()
-    }
-}
-
-
-const chartConfig = (nodesUsage: UserNodeUsagesResponse) => {
-    const numberOfNodes = nodesUsage.node_usages.length;
-    const config: Record<string, any> = {
-        views: {
-            label: "Page Views",
-        }
-    }
-    const colorRangeInfo = {
-        colorStart: 0,
-        colorEnd: 1,
-        useEndAsStart: false,
-    };
-    const colors = interpolateColors(numberOfNodes, interpolateRainbow, colorRangeInfo);
-    nodesUsage.node_usages.forEach((node, i) => {
-        config[node.node_name] = { label: node.node_name, color: colors[i] };
-    })
-    return config satisfies ChartConfig;
-}
 
 export const UserNodesUsageWidget: FC<UserNodesUsageWidgetProps> = ({
     user,
@@ -70,8 +22,8 @@ export const UserNodesUsageWidget: FC<UserNodesUsageWidgetProps> = ({
     const [timeRange, setTimeRange] = useState("90d")
     // const { start, end } = useStartEndDateFromNow(timeRange as ChartDateInterval);
     // const { data, isLoading } = useUserNodeUsagesQuery({ username: user.username, start, end })
-    const chartData = transformData(mockData);
-    const config = chartConfig(mockData);
+    const chartData = useTransformData(mockData);
+    const config = useChartConfig(mockData);
 
     return (
         <SectionWidget
@@ -124,7 +76,7 @@ export const UserNodesUsageWidget: FC<UserNodesUsageWidgetProps> = ({
                                 <stop
                                     offset="5%"
                                     stopColor={config[node.node_name].color}
-                                    stopOpacity={0.8}
+                                    stopOpacity={0.9}
                                 />
                                 <stop
                                     offset="95%"
