+++ dashboard/src/modules/users/components/widgets/user-nodes-usage-widget/types.ts
@@ -1,0 +1,12 @@
+import { UserType } from "@marzneshin/modules/users";
+
+export interface NodesUsage {
+    datetime: Date;
+    nodes: Record<string, number>;
+}
+
+export interface UserNodesUsageWidgetProps {
+    user: UserType;
+}
+
+export type ChartDateInterval = '90d' | '30d' | '7d';
